:hammer: code base major restructure to improve project growth
