+++ app/setup.py
@@ -1,7 +1,8 @@
-from app import app
-import cloudinary as cloudinary
 import os
+from app import app
 from flask_pymongo import PyMongo
+import cloudinary as cloudinary
+from datetime import date
 
 # cloudinary config
 cloudinary.config(
@@ -17,7 +18,15 @@ app.config["MONGO_URI"] = os.environ.get('HC_MONGO_URI')
 # secret key
 app.config["SECRET_KEY"] = os.environ.get('SESSION_SECRET')
 
-# constant variables
+######################
+# CONSTANT VARIABLES #
+######################
+
+# Today date String
+TODAY_STR = date.today().strftime("%d/%m/%Y")
+# Database collections
 MONGO = PyMongo(app)
 DB_RECIPES = MONGO.db.recipes
 DB_USERS = MONGO.db.users
+DB_INGREDIENTS = MONGO.db.ingredients
+DB_METHODS = MONGO.db.methods
