+++ app/templates/edit-recipe.html
@@ -272,7 +272,5 @@
             </div>
         </div>
     </div>
-    </div>
-
 
 {% endblock %}
